Harmonize pipeline.parameters init/default values/checks for fmri-preprocessing
