+++ clinica/pipelines/fmri_preprocessing/fmri_preprocessing_cli.py
@@ -16,7 +16,9 @@ class fMRIPreprocessingCLI(ce.CmdParser):
 
     def define_options(self):
         """Define the sub-command arguments."""
+        from .fmri_preprocessing_utils import get_pipeline_parameters
         from clinica.engine.cmdparser import PIPELINE_CATEGORIES
+        parameters = get_pipeline_parameters()
         # Clinica compulsory arguments (e.g. BIDS, CAPS, group_id)
         clinica_comp = self._args.add_argument_group(PIPELINE_CATEGORIES['CLINICA_COMPULSORY'])
         clinica_comp.add_argument("bids_directory",
@@ -28,22 +30,24 @@ class fMRIPreprocessingCLI(ce.CmdParser):
         optional = self._args.add_argument_group(PIPELINE_CATEGORIES['OPTIONAL'])
         optional.add_argument("-fwhm", "--full_width_at_half_maximum",
                               metavar="FWHM",
-                              nargs=3, type=int, default=[8, 8, 8],
+                              nargs=3, type=int, default=parameters['full_width_at_half_maximum'],
                               help="A list of integers specifying the full width at half maximum (FWHM) filter size "
                                    "in the three directions x y z, in millimeters, used to smooth the images "
-                                   "(default: --full_width_at_half_maximum 8 8 8).")
-        optional.add_argument("-t1s", "--t1_native_space", action='store_true',
+                                   "(default: --full_width_at_half_maximum %s)." %
+                                   self.list_to_string(parameters['full_width_at_half_maximum']))
+        optional.add_argument("-t1s", "--t1_native_space",
+                              action='store_true', default=parameters['t1_native_space'],
                               help="When specified, the corrected fMRI data registered to the native T1 space "
                                    "of each subject are saved. This might be useful when dealing with "
                                    "parcellations that are in the native T1 space.")
         optional.add_argument("-fsbm", "--freesurfer_brain_mask",
-                              action='store_true',
+                              action='store_true', default=parameters['freesurfer_brain_mask'],
                               help="When specified, the pipeline uses the brain extraction generated by FreeSurfer "
                                    "during the t1-freesurfer pipeline. Otherwise, it computes it using the "
                                    "SPM Registration tool followed by empirically-estimated operations "
                                    "performed by FSL.")
         optional.add_argument("-u", "--unwarping",
-                              action='store_true',
+                              action='store_true', default=parameters['unwarping'],
                               help="When specified, the unwarping correction step is executed. This will add "
                                    "SPM's Unwarping to the Realign step (magnitude and phasediff files are necessary "
                                    "in the BIDS directory).")
@@ -59,20 +63,23 @@ class fMRIPreprocessingCLI(ce.CmdParser):
         """Run the pipeline with defined args."""
         from networkx import Graph
         from .fmri_preprocessing_pipeline import fMRIPreprocessing
+        from .fmri_preprocessing_utils import get_pipeline_parameters
         from clinica.utils.ux import print_end_pipeline, print_crash_files_and_exit
 
+        parameters = get_pipeline_parameters(
+            full_width_at_half_maximum=args.full_width_at_half_maximum,
+            t1_native_space=args.t1_native_space,
+            freesurfer_brain_mask=args.freesurfer_brain_mask,
+            unwarping=args.unwarping
+        )
         pipeline = fMRIPreprocessing(
             bids_directory=self.absolute_path(args.bids_directory),
             caps_directory=self.absolute_path(args.caps_directory),
             tsv_file=self.absolute_path(args.subjects_sessions_tsv),
-            base_dir=self.absolute_path(args.working_directory)
+            base_dir=self.absolute_path(args.working_directory),
+            parameters=parameters,
+            name=self.name
         )
-        pipeline.parameters = {
-            'full_width_at_half_maximum' : args.full_width_at_half_maximum,
-            't1_native_space'            : args.t1_native_space,
-            'freesurfer_brain_mask'      : args.freesurfer_brain_mask,
-            'unwarping'                  : args.unwarping
-        }
 
         if args.n_procs:
             exec_pipeline = pipeline.run(plugin='MultiProc',
